AI/Petnamesgenerator/requirements.txt
langchain
streamlit
python-dotenv